# docs/docs/api/openapi/ui-backend-api.yaml
openapi: 3.1.0
info:
  title: CAIPE UI Backend API
  version: 1.0.0
  description: |
    Backend-for-Frontend API for the CAIPE UI (Next.js App Router). Routes live under `/api/*` on the UI origin.
    Most endpoints require a NextAuth session cookie. Admin vs admin-view vs authenticated-only behavior matches
    the markdown API docs in `docs/docs/api/`.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api
    description: UI Backend API routes
security:
  - SessionCookie: []
...
paths:
...
  /settings:
    get:
      operationId: getUserSettings
      summary: Full user settings document
      tags: [settings]
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessUserSettings"
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: putUserSettings
      summary: Partial update nested preferences/notifications/defaults
      tags: [settings]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsPatch"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessUserSettings"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: |
        NextAuth.js session cookie. In production with Secure cookies the name may be
        `__Secure-next-auth.session-token`. Send the session cookie from the browser; API clients must obtain a session via OIDC first.
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
...
  schemas:
    ApiErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
        code:
          type: string
...
    UserSettingsPreferences:
      type: object
      additionalProperties: true
      properties:
        theme: { type: string }
        gradient_theme: { type: string }
        font_family: { type: string }
        font_size: { type: string }
        sidebar_collapsed: { type: boolean }
        context_panel_visible: { type: boolean }
        debug_mode: { type: boolean }
        code_theme: { type: string }
    UserSettingsNotifications:
      type: object
      additionalProperties: true
      properties:
        email_enabled: { type: boolean }
        in_app_enabled: { type: boolean }
        conversation_shared: { type: boolean }
        weekly_summary: { type: boolean }
    UserSettingsDefaults:
      type: object
      additionalProperties: true
      properties:
        default_model: { type: string }
        default_agent_mode: { type: string }
        auto_title_conversations: { type: boolean }
    UserSettings:
      type: object
      properties:
        _id: { type: string }
        user_id: { type: string, format: email }
        preferences:
          $ref: "#/components/schemas/UserSettingsPreferences"
        notifications:
          $ref: "#/components/schemas/UserSettingsNotifications"
        defaults:
          $ref: "#/components/schemas/UserSettingsDefaults"
        updated_at: { type: string, format: date-time }
    SuccessUserSettings:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data:
          $ref: "#/components/schemas/UserSettings"
    UserSettingsPatch:
      type: object
      properties:
        preferences:
          $ref: "#/components/schemas/UserSettingsPreferences"
        notifications:
          $ref: "#/components/schemas/UserSettingsNotifications"
        defaults:
          $ref: "#/components/schemas/UserSettingsDefaults"
...